refactor(segmentation): split segmentation threshold effect file
- Move threshold pipeline and parameters to a dedicated file
- Move threshold opacity blinker to a dedicated file
- Rename hatch parameter
- Move outline hatch parameters to internal pipeline values
- Remove uninteresting comments in tests
